+++ plugins/ValheimServerManager.Bootstrap/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("ValheimServerManager.Bootstrap.Tests")]
-[assembly: InternalsVisibleTo("ValheimServerManager.Plugin.Tests")]
-[assembly: InternalsVisibleTo("ValheimServerManagerRuntimeUpdater")]